+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/YearEnd/TerminatedEmployeeAndBeneficiaryDataResponseDto.cs
@@ -11,7 +11,7 @@ public sealed record TerminatedEmployeeAndBeneficiaryDataResponseDto
     public required decimal VestedBalance { get; set; }
     public required DateOnly? DateTerm { get; set; }
     public required decimal YtdPsHours { get; set; }
-    public required int VestedPercent { get; set; }
+    public required decimal VestedPercent { get; set; }
     public required int? Age { get; set; }
     public required byte? EnrollmentCode { get; set; }
